forward declare drawFrameCallback with the type
originally the drawFrameCallback function was just placed on top of the
main function and the compiler performed typecheck in the line where
it's passed to the winRegisterDrawFrameCallback function. now instead we
declare it with a typedef and define it below the main function.

this is more robust way of making sure that the function type is what
the callback expects.